\name{findUnlabelSamples}
\alias{findUnlabelSamples}



\title{
Find unlabeleded samples
}


\description{
The unlabeled samples are extracted from the same set of transcripts with consensus motif (e.g. RRACH) of confident positive samples, but exclude the confident positive samples.
}


\usage{
findUnlabelSamples(cDNAID, motifPos, posSamples, ...)
}

                  
\arguments{
   \item{cDNAID}{A character vector representing the transcripts IDs.}
   \item{motifPos}{A list containing the position of the consensus motif for each transcript.}
   \item{posSamples}{A character vetor specified the positive samples, which can be generated by function "findConfidentPosSamples".}
   \item{\dots}{Further paramters used for find confident positive samples.}
}



\value{
  A list containing two elements: the positive samples and its transcript IDs.
}


\author{
  Jingjing Zhai, Chuang Ma, Jie Song
}


\examples{
\dontrun{

  cDNA <- '/home/malab14/tair10_cDNA.fa'
  GTF <- '/home/malab14/Arabidopsis.gtf'
  #Convert genomic position to cDNA position
  peaks <- G2T(bedPos = cmrMat, GTF = GTF)
  ##"cmrMat" can be generated by function "CMRCalling".
  ##Search consensus motif in cDNA sequence
  motifPos <- searchMotifPos(RNAseq = cDNA)
  posSamples <- findConfidentPosSamples(peaks = peaks, motifPos = motifPos)
  unlabelSamples <- findUnlabelSamples(cDNAID = posSamples$cDNAID, 
                                       motifPos = motifPos, 
                                       posSamples = posSamples$positives)
  
}
}



