\name{findUnlabelSamples}
\alias{findUnlabelSamples}



\title{
Find unlabeleded samples
}


\description{
The unlabeled samples are extracted from the same set of transcripts with consensus motif (e.g. RRACH) of confident positive samples, but exclude the confident positive samples.
}


\usage{
findUnlabelSamples(cDNAID, motifPos, posSamples, ...)
}

                  
\arguments{
   \item{cDNAID}{A character vector representing the transcripts IDs.}
   \item{motifPos}{A list containing the position of the consensus motif for each transcript.}
   \item{posSamples}{A character vetor specified the positive samples, which can be generated by function "findConfidentPosSamples".}
   \item{\dots}{Further paramters used for find confident positive samples.}
}



\value{
  A list containing two elements: the positive samples and its transcript IDs.
}


\author{
  Jingjing Zhai, Chuang Ma, Jie Song
}


\examples{
\dontrun{

  ##Load data
  input.bam <- system.file("extdata/chr1_input_test.bam", package = "PEA")
  RIP.bam <- system.file("extdata/chr1_RIP_test.bam", package = "PEA")
  refGenome <- system.file("extdata/chromosome1.fa", package = "PEA")
  cDNA <- system.file("extdata/chr1_cdna.fa", package = "PEA")
  GTF <- system.file("extdata/chromosome1.gtf", package = "PEA")

  ##m6A peak calling using sliding window-based method
  cmrMat <- CMRCalling(CMR = "m6A", IPBAM = RIP.bam, inputBAM = input.bam,
                       method = "SlidingWindow", mappedInput = 17472, 
                       mappedRIP = 20072, refGenome = refGenome)

  
  
  #Convert genomic position to cDNA position
  peaks <- G2T(bedPos = cmrMat, GTF = GTF)
  
  ##Search consensus motif in cDNA sequence
  motifPos <- searchMotifPos(sequence = cDNA)
  
  ##Find confident positive samples
  posSamples <- findConfidentPosSamples(peaks = peaks, motifPos = motifPos)
  
  
  ##Find unlabel samples
  unlabelSamples <- findUnlabelSamples(cDNAID = posSamples$cDNAID, 
                                       motifPos = motifPos,
                                       posSamples = posSamples$positives)

  
}
}



